+++ client/WorldView.h
@@ -50,7 +50,7 @@ public:
 
     void updatePositions(std::map<uint8_t, std::pair<float, float>>& positionMap);
     void updateAngles(std::map<uint8_t, int16_t>& angles);
-    void render(uint8_t iteration);
+    void render(size_t iteration);
 
     ~WorldView();
