+++ client/anim/Renderizable.h
@@ -21,7 +21,7 @@ public:
 
     virtual void updatePosition(float x, float y);
     void updateAngle(int16_t newAngle);
-    virtual void render(Camera &camera, uint8_t iteration) = 0;
+    virtual void render(Camera &camera, size_t iteration) = 0;
 
     virtual ~Renderizable();
 };
